Module price_feed::util

Enums

Error
Ethabi errors

ParamType
Function and event param types.

Token
Ethereum ABI params.

Functions

decode
Decodes ABI compliant vector of bytes into vector of tokens described by types param.

slice_data
Converts a vector of bytes with len equal n * 32, to a vector of slices.

Type Definitions

Int
ABI signed integer.

Uint
ABI unsigned integer.

Word
ABI word.